Support archived versions.
Since rubygems.org reports only the most recent version for each gem,
keep a local list of the versions theat were already downloaded. In the
case we package one version and upstream already released a more recent
one, we do not want to lose the tarball for the packaged version (since
for Debian it will be *the* orig tarball).